mqtt/logger/requirements.txt
paho-mqtt
pymodbus>=3.0.0